Added lots of page content and some styles
Borrowed a lot of stuff from the actual wikipedia page.
The general thought on this was kinda twofold. Some stuff can only be written one way, like the top right links (Arts etc), and the main content would either have to be laboriously written out to look similar, or be generic filler and look bad.